@@ -1,3 +1,3 @@
 class GardenCropSerializer < ApplicationSerializer
-  attributes :_id, :guide, :sowed, :stage, :quantity
+  attributes :_id, :guide, :crop, :sowed, :stage, :quantity
 end
